+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/ChinookAPI.java
@@ -1,4 +1,4 @@
-package is.codion.chinook.domain;
+package is.codion.chinook.domain.api;
 
 import static is.codion.framework.domain.DomainType.domainType;